+++ src/index.ts
@@ -1,3 +1,3 @@
-export { Router } from './Router';
 export * from './components';
 export * from './types';
+export { Router } from './Router';