+++ src/types/index.ts
@@ -1,5 +1,5 @@
 export type { RouteConfig } from './RouteConfig';
+export { RouteError, NotFoundRouteError } from './RouteError';
+export { RouteBeginEvent, RouteDoneEvent, RouteErrorEvent } from './RouteEvent';
 export type { RouteInfo } from './RouteInfo';
 export type { RouterConfig } from './RouterConfig';
-export { RouteError, NotFoundRouteError } from './RouteError';
-export * from './RouteEvent';